refactor(autocomplete): remove AutocompleteBase and AutocompleteMultiColumn exports and stories
and add both under the broader Autocomplete component
