+++ packages/components/src/index.ts
@@ -8,10 +8,6 @@ export * from "./core/Alert";
 export { default as Alert } from "./core/Alert";
 export * from "./core/Autocomplete";
 export { default as Autocomplete } from "./core/Autocomplete";
-export * from "./core/AutocompleteBase";
-export { default as AutocompleteBase } from "./core/AutocompleteBase";
-export * from "./core/AutocompleteMultiColumn";
-export { default as AutocompleteMultiColumn } from "./core/AutocompleteMultiColumn";
 export * from "./core/Banner";
 export { default as Banner } from "./core/Banner";
 export * from "./core/Button";
